Explicit cast of -1 to UINT, to silence GCC warnings, but still adhere to the strange win32 API.
git-svn-id: svn://svn.code.sf.net/p/tigervnc/code/trunk@3376 3789f03b-4d11-0410-bbf8-ca57d06f2519

+++ win/vncviewer/CConn.cxx
@@ -436,7 +436,7 @@ CConn::setColourMapEntries(int first, int count, U16* rgbs) {
 void
 CConn::bell() {
   if (options.acceptBell)
-    MessageBeep(-1);
+    MessageBeep((UINT)-1);
 }
